backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
pyfao56==1.3.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2 
python-dotenv==1.0.0
Pillow==10.0.0
opencv-python==4.8.0.76
Werkzeug==2.3.7
keras==2.13.1
tensorflow==2.13.0
Flask-SQLAlchemy==3.0.3
APScheduler==3.11.0